requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# HTTP requests
requests>=2.31.0

# Environment management
python-dotenv>=1.0.0

# Logging and utilities
pathlib2>=2.3.0

# Data validation (optional but recommended)
pydantic>=2.0.0

# Testing (for future development)
pytest>=7.0.0
pytest-cov>=4.0.0

# Analysis and visualization (for future notebooks/analysis)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
jupyter>=1.0.0

# Development tools
black>=23.0.0
flake8>=6.0.0
